Address PR comments: move useSingleInstanceIdsData hook to EnsureSingleInstance
